Fix #10; Show plagiarised source text and add improvements
To show the plagiarised source text I had to delete dublicates in similarities_src and sort the array from lowest to highest. This ensures that you can readd the symbols and not stop the coloring process when printing the result. I also fixed an issue where the j loop value in the compare function would become less than 0