Remove all symbols and keep 'æøå'
This solution keeps 'æøå' because we’re only comparing sections of the character.
strchr basically takes the first parameter string value and loops through all the characters. It then compares those characters to the one we provided as the second parameter value which we also got from looping through a string (the text). When you get to the ‘æøå’ characters in a loop, it essentially needs to divide it into two sections, since it’s size of 32-bit is twice as big as the normal 16-bit character. This means that “æ” fx. needs two loops to get through, where the loops represents the first and second half of the character. This means that the function compares the half of the character instead of the whole character.