add word conjugation checker
The conjugation function checks if the words are the same irregardless of the conjugation. So if we're comparing a word with it's plural form then it would flag it as being similar. The function compares the characters in both strings and increments the pointer position until the characters doesn't equal to each other. If the word was the plural form of another word then one string would be empty and the other would contain the plural suffix